+++ en.js
@@ -959,7 +959,9 @@ export default {
   'workshop.scenesManager': 'Scenes Manager',
   'common.timelines': 'Timelines',
   // added 2019-05-06
-  'universe.authors.title.meta': '{title} authors'
+  'universe.authors.title.meta': '{title} authors',
+  // added 2019-06-27
+  'workshop.collaborators.addRoleForLang': 'Add {role} for {language}'
 };
 
 /* eslint-enable max-len */